Engine - frontend communication
A lot of refactoring was done. An clean interface between recognition
engine and graphics frontend was added. Interfaces have been specified
using classes in frontend.py.

Extended straight_line_detector(). It returns a qualifier in addition
to the detection decision. This qualifier tells if the line is vertical,
horizontal or diagonal.