Ray angles can be quantized
When changing a ray orientation using the handle, the angles allowed
are limited to 8 or 16 values, depending on the distance between
handle a mouse cursor. If the cursor is far enough (around 140 pixels
now), no quantization is applied.

This permits perfectly parallel rays (with each other and with the
baseline)